[jaepil] feat: Add more training hyperparameters to compare model performance
Decrease lr, add warmup_steps, and higher the batch size. Also, use EM as the metric for the best model.

Have to figure out what is the cause of the sudden jump in performance.

Previously, EM was 0.0 and F1 was 3.577% but now it is 14.58% and F1 is 23.14%, which is a big jump.

I changed many arguments at once:
- lr
- warmup_steps
- per_device_train_batch_size
- metric_for_best_model & greater_is_better

So it is not exactly clear what contributed the most to improve the model performance.

I suspect setting EM as the loss function. I'll try to comment them out and do the experiment again.